Add files via upload
Structure:

The CustomExpansionTile widget is designed to resemble an expansion tile with a title, trailing icons (in the form of a GridView), and expandable/collapsible content.
It is implemented as a stateful widget with a mutable _isExpanded variable to track the expansion state.
Usage:

You can use this custom expansion tile in your Flutter app by providing a title, a list of children (content to be displayed when expanded), and an optional list of trailingGridView (icons displayed on the right side).
Functionality:

When the user taps on the InkWell (which wraps the title and trailing icons), the _isExpanded state is toggled, expanding or collapsing the content accordingly.
The GridView is used to display the trailing icons, and it's configured with fixed cross-axis count, spacing, and physics to prevent scrolling.
Styling:

The widget is styled using a Container with a specific decoration for a rounded appearance and padding for spacing.
A Divider is shown when the tile is expanded to separate the trailing icons from the content.
Differences from Flutter's ExpansionTile:

The custom expansion tile provides a bit more flexibility in terms of layout and styling compared to Flutter's built-in ExpansionTile. It allows you to have a custom arrangement of title, trailing icons, and content.
Flutter's ExpansionTile typically comes with built-in animations and follows the Material Design guidelines for consistency. The custom implementation might not have the same level of built-in animations.
The custom expansion tile offers more control over the appearance, allowing you to customize the styling, layout, and behavior.
Benefits:

The custom expansion tile is beneficial when you need a specific layout or styling that might not be achievable using the default ExpansionTile.
It allows you to include a dynamic number of trailing icons in the form of a GridView, providing a visually appealing way to present additional actions or information.
The custom implementation can be useful in scenarios where you want to tailor the appearance and behavior to meet specific design requirements.
Remember that the choice between using a custom implementation like this and the Flutter-provided ExpansionTile depends on your specific use case, design preferences, and the level of customization and control you require.